added Entities for each model to separate business logic from data access. Added configuration classes(not implemented)
