+++ ProjectManagementAPI/Configurations/ProjectTasksConfiguration.cs
@@ -1,0 +1,13 @@
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using ProjectManagementAPI.Entities;
+
+namespace ProjectManagementAPI.Configurations;
+
+public class ProjectTasksConfiguration : IEntityTypeConfiguration<ProjectTaskEntity>
+{
+    public void Configure(EntityTypeBuilder<ProjectTaskEntity> builder)
+    {
+        throw new NotImplementedException();
+    }
+}
