+++ ProjectManagementAPI/Configurations/ProjectsConfiguration.cs
@@ -1,0 +1,13 @@
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using ProjectManagementAPI.Entities;
+
+namespace ProjectManagementAPI.Configurations;
+
+public class ProjectsConfiguration : IEntityTypeConfiguration<ProjectEntity>
+{
+    public void Configure(EntityTypeBuilder<ProjectEntity> builder)
+    {
+        throw new NotImplementedException();
+    }
+}
