+++ ProjectManagementAPI/Configurations/ProjectsUsersConfiguration.cs
@@ -1,0 +1,13 @@
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using ProjectManagementAPI.Models;
+
+namespace ProjectManagementAPI.Configurations;
+
+public class ProjectsUsersConfiguration : IEntityTypeConfiguration<ProjectUser>
+{
+    public void Configure(EntityTypeBuilder<ProjectUser> builder)
+    {
+        throw new NotImplementedException();
+    }
+}
