+++ ProjectManagementAPI/Configurations/UsersConfiguration.cs
@@ -1,0 +1,13 @@
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using ProjectManagementAPI.Entities;
+
+namespace ProjectManagementAPI.Configurations;
+
+public class UsersConfiguration : IEntityTypeConfiguration<UserEntity>
+{
+    public void Configure(EntityTypeBuilder<UserEntity> builder)
+    {
+        throw new NotImplementedException();
+    }
+}
