+++ ProjectManagementAPI/Entities/UserEntity.cs
@@ -1,11 +1,14 @@
-using System.ComponentModel.DataAnnotations;
+using ProjectManagementAPI.Models;
 
-namespace ProjectManagementAPI.Models;
+namespace ProjectManagementAPI.Entities;
 
-public class User
+public class UserEntity
 {
     public Guid Id { get; set; }
+    
     public string FirstName { get; set; } = string.Empty;
+    
     public string LastName { get; set; } = string.Empty;
+    
     public List<ProjectUser> ProjectUsers { get; set; } = new();
 }
